fix: Update datetime only for the focused window
Currently, when an app is activated or focused, the datetime
of all its windows is updated. This causes all app windows
to be sorted to the front in the window switcher, making it
confusing and harder to use.

Now, only the focused window’s datetime is updated, providing
a more intuitive experience and aligning with Windows OS behavior.

Partially fixes #479